Feature/retoques func cliente (#11)
* Cambio de título al loguear/desloguear cliente.

* Agregados algunos clientes de test más. Cambiados nombres de los clientes de test para identificarlos fácilmente por su subtipo de cliente.

* Arreglado un bug en el método aplicarBeneficioServicio de DeLaCasa; si el cliente gastaba menos de 500, iba a devolver un número negativo como precio en vez de devolver 0.

* Limpieza del monto entre servicios de clientes (se pone en 0 en la GUI).

* Reintegración de stock cuando se elimina un pedido.

* Manejo de excepciones y actualización de pedidos cuando no hay stock para confirmar todos los pedidos.

* Manejo de excepciones: agregados los CA de eliminarPedido: 'debe identificarse' si el cliente está en null y 'debe seleccionar un pedido' si el cliente está logeuado pero no seleccionó pedidos para eliminar. BUGFIX IMPORTANTE: En ControladorDispositivo, al iniciar sesion, se guarda el cliente (del servicio) en una variable cliente. Pero esta variable no se ponía en null al terminar el servicio, por lo que al hacer login y luego logout, algunas acciones que requerían usuario logueado no mostraban el error correspondiente. Arreglado al incluír cliente = null; en terminarServicio().

* Cambiado el mensaje de eliminar pedido en PedidoEnProceso para estar más acorde al mensaje que pide el CA.

* Usando el EstadoPedido.TipoDeEstado para comparaciones en vez de usar strings (buena práctica... creo... si tenemos un enum, mejor usarlo).

* Implementados y arreglados CAs 3 y 5 (pedidos que se borran al confirmar por falta de stock y pedidos no confirmados que se borran si alguien más confirma y deja sin stock). stockDeItemsSinConfirmar implementado en fachada y subsistema servicio, es llamado desde el controlador de dispositivo cuando se reciben eventos de cambio de estado de pedidos.

+++ src/Modelo/EstadosDePedido/Pedido.java
@@ -40,8 +40,16 @@ public class Pedido {
         return gestor;
     }
     
-    public String getEstado(){
+    public EstadoPedido getEstado(){
+        return estado;
+    }
+    
+    public String getEstadoTexto(){
         return estado.toString();
+    }
+    
+    public EstadoPedido.TipoDeEstado getTipoDeEstado(){
+        return estado.getTipo();
     }
     
     protected void setEstado(EstadoPedido e){
@@ -54,12 +62,15 @@ public class Pedido {
     }
     
     protected void hacerEliminarPedido(){
+        if (estado.getTipo() != EstadoPedido.TipoDeEstado.NO_CONFIRMADO) {
+            this.item.devolverStockDeInsumos();            
+            this.getItem().getUnidadProcesadora().eliminarPedido(this);
+        }
         servicio.removerPedido(this);
         this.servicio = null;
-        this.getItem().getUnidadProcesadora().eliminarPedido(this);
     }
     
-    protected void hacerConfirmarPedido(){
+    protected void hacerConfirmarPedido() throws PolloException {
         this.getItem().getUnidadProcesadora().agregarPedido(this);
     }
